requirements.txt
Pillow>=10.0
pillow-heif>=0.13
requests>=2.28
google-genai>=1.0
python-dotenv>=1.0
